pyproject.toml
[project]
name = "discreet-ai-coding-assistant"
version = "1.0.0"
description = "A discreet AI coding assistant that remains invisible during screen sharing"
authors = [
    {name = "AI Coding Assistant Team"}
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.70.0",
    "psutil>=5.9.0",
    "pynput>=1.7.6",
]

[project.optional-dependencies]
gui = [
    "PyQt5>=5.15.0",
]
windows = [
    "pywin32>=302",
]
macos = [
    "pyobjc-core>=9.0.1",
    "pyobjc-framework-Cocoa>=9.0.1",
    "pyobjc-framework-Quartz>=9.0.1",
]

[tool.setuptools]
packages = ["ai_assistant"]
